TarScript class

Stores all the required commands used for pipeline building.

Methods

Public methods

TarScript$new()

TarScript$build()

TarScript$update()

TarScript$print()

TarScript$deparse_script()

Method new()

Create a new TarScript object.

Usage

TarScript$new(
  package = c("targets", "tarchetypes"),
  global = NULL,
  option = NULL,
  targets = NULL,
  pipeline = NULL
)

Arguments

package

Character vector. The package need to be included in the script. Usually only "targets" and "tarchetypes" only. Set it NULL if not needed. And repetitive packages are removed.

global

A character vector. Global variables or functions used in the script. For example, there might be a source() call to prepare functions. character used for better formatting. No unique() is applied.

option

A list. Will be arguments of tar_option_set(). Set it NULL if not needed.

targets

A character vector. Defined targets out of the pipeline list. character used for better formatting. No unique() is applied.

pipeline

A list of expressions, each of which is a call to tar_target() or its related. This cannot be NULL when build.

Method build()

Build the targets script. This will update the file "_targets.R" from current project by default.

Usage

TarScript$build(path = NULL, ..., styler = TRUE)

Arguments

path

A connection, or a character string naming the file to print the script. Set to "" to print to the standard output. The default is "_targets.R".

...

For future expansion use and must be empty.

styler

A logical indicating if styler should be called to make the generated file nicer to read. Default is TRUE.

Method update()

Update a part of the script.

Usage

TarScript$update(step, codes, append = TRUE)

Arguments

step

A character of the updating step name.

codes

The codes to update.

append

A logical value indicating if the codes are appended to the old step (TRUE) or replaced (FALSE).

Method print()

Print object more informative.

Usage

TarScript$print(...)

Arguments

...

For future expansion use and must be empty.

Method deparse_script()

Deparse the whole script to a character vector. Each element of it is a code line. Note this method does not return the object itself, so it cannot be chained.

Usage

TarScript$deparse_script()