Fetch datasets from given database

Based on a template query file and "where clause" configuration, datasets are extracted from a given database.

fetch(query_file, config_where = NULL, dsn = "iquizoo-v3", encoding = "utf-8")

fetch_single_game(query_file, game_id, config_where = NULL, ...)

Arguments

query_file
File name of sql query

config_where
Configuration of "where-clause" of the sql query. Can be a list (mostly from the config.yml file) or data.frame.

dsn
The data source name of an ODBC database connector. See odbc::dbConnect() for more information.

encoding
Encoding to be assumed for input strings. Default to "UTF-8".

game_id
The identifier of the game to fetch datasets from.

...
Other arguments passed to fetch().

Value

A tibble of the fetched data.

Functions

fetch: Default usage of fetch.

fetch_single_game: A special case to fetch datasets from a single game.

Author

Liang Zhang