Setup Workflow

Create a basic infrastructure for iquizoo data analysis.

init(wizard = interactive(), schema = "scores", separate = NULL)

Arguments

wizard
A logical value indicating whether to use interactive operations. If TRUE, menus will be shown interactively.

schema
The name of workflow to be used. Different name means different targets pipelines.

separate
Should the database downloading be separated into branches based on games?

Value

NULL (invisibly). This function is called for its side-effect.

Details

These steps are done in order:

Prepare configuration file if not found. Configuration file is named as "config.yml", which is basically used to set the where-clause for database queries (SQL).

Prepare database query files (SQL). The major part of the whole work is just to download data from database, so for now, up to four query files are used. Note they are all "templates" only, cannot be used directly.

Prepare "_targets.R". Package targets parses commands from this R script, which plays a role as a workflow blueprint. In this step, users will be asked to choose which "schema" to use if wizard is TRUE, typically when in interactive mode.

Add "_targets" to file .gitignore if found.

Manually check configuration file "config.yml".