Insert where configurations

Insert new where configurations into the existed one. This can be useful in many projects doing minor changes to current where configurations.

insert_where(old, ...)

# S3 method for `NULL`
insert_where(old, ...)

# S3 method for list
insert_where(old, ..., replace = TRUE)

# S3 method for data.frame
insert_where(old, ..., replace = TRUE)

insert_where_single_game(old, game_id)

Arguments

old

The old where configuration.

...

The new where element. Can be of list or data.frame class. Note they must have name of "table". What's more, to make it work properly in the whole project, the names should be no other than c("table", "field", "operator", "values").

replace

Replace the element with the same table value or not. Default is set to replace, set FALSE to add it anyway (not recommended, might cause error in further analysis).

game_id

The identifier of the game to set the limitation.

Value

A where configuration of the same format with the old one.

Details

This function is intended to modify an existed where configuration. If the existed one is_empty(), the added will be returned as a list().

Functions

insert_where_single_game: Limit old where config to one single game.

Author

Liang Zhang