Generate a set of targets for wrangling and pre-processing raw data

This target factory is the main part of the tar_prep_iquizoo function. It wrangles the raw data into a tidy format and calculates indices based on the parsed data.

Usage

tar_prep_raw(
  contents,
  action_raw_data = c("parse", "preproc"),
  name_data = "raw_data",
  name_parsed = "raw_data_parsed",
  name_indices = "indices"
)

Arguments

contents

The contents structure used as the configuration of data fetching.

action_raw_data

The action to be taken on the fetched raw data.

name_data

The name of the raw data target.

name_parsed

The name of the parsed data target.

name_indices

The name of the indices target.

Value

A list of target objects.