Generate a set of targets for fetching data

This target factory is the main part of the tar_prep_iquizoo function. It fetches the raw data and scores for each project and task/game combination.

Usage

tar_fetch_data(
  contents,
  templates = setup_templates(),
  what = c("raw_data", "scores"),
  check_progress = TRUE
)

Arguments

contents

The contents structure used as the configuration of data fetching.

templates

The SQL template files used to fetch data. See setup_templates() for details.

what

What to fetch.

check_progress

Whether to check the progress hash. If set as TRUE, Before fetching the data, the progress hash objects named as progress_hash_{project_id} will be depended on, which are typically generated by tar_prep_proj(). If the projects are finalized, set this argument as FALSE.

Value

A list of target objects.