Generate a set of targets for pre-processing of iQuizoo data

This target factory prepares a set of target objects used to fetch data from iQuizoo database, separated into static branches so that each is for a specific project and task/game combination. Further pre-processing on the fetched data can also be added if requested.

Usage

tar_prep_iquizoo(
  params,
  ...,
  contents = NULL,
  what = c("all", "raw_data", "scores"),
  action_raw_data = c("all", "parse", "none"),
  templates = setup_templates(),
  check_progress = TRUE
)

Arguments

params

A data.frame or list contains the parameters to be bound to the query. Default templates require specifying organization_name and project_name, in that order. If contents template is specified without any parameters, set it as empty vector or NULL. If contents argument is specified, this argument is omitted.

...

For future usage. Should be empty.

contents

The contents structure used as the configuration of data fetching. It is typically automatically fetched from database based on the contents template in templates. If not NULL, it will be used directly and ignore that specified in templates. Note contents should at least contains project_id and game_id names.

what

What to fetch. There are basically two types of data, i.e., raw data and scores. The former is the logged raw data for each trial of the tasks/games, while the latter is the scores calculated by iQuizoo server. If set as "all", both raw data and scores will be fetched. Further actions on the fetched raw data can be specified by action_raw_data.

action_raw_data

The action to be taken on the fetched raw data. There are two consecutive actions, i.e., wrangling and pre-processing. The former will parse the raw data into a tidy format, while the latter will calculate indices based on the parsed data. If set as "all", both wrangling and pre-processing will be done. If set as "parse", only wrangling will be done. If set as "none", neither will be done. If what is "scores", this argument will be ignored.

templates

The SQL template files used to fetch data. See setup_templates() for details.

check_progress

Whether to check the progress hash. Set it as FALSE if the project is finalized.

Value

A list of target objects.